+++ ocean_provider/utils/basics.py
@@ -13,8 +13,6 @@ from hexbytes import HexBytes
 from ocean_provider.config import Config
 from ocean_provider.http_provider import CustomHTTPProvider
 from ocean_provider.utils.asset import Asset
-from ocean_provider.utils.datatoken import get_dt_contract
-from requests_testadapter import Resp
 from web3 import WebsocketProvider
 from web3.main import Web3
 
@@ -53,15 +51,6 @@ def get_provider_wallet(web3: Optional[Web3] = None):
         )
 
     return wallet
-
-
-def get_datatoken_minter(datatoken_address):
-    """
-    :return: Eth account address of the Datatoken minter
-    """
-    dt = get_dt_contract(get_web3(), datatoken_address)
-    publisher = dt.caller.minter()
-    return publisher
 
 
 def get_artifacts_path():
